move some enums to GTA .h files
