+++ source/StdInc.h
@@ -376,22 +376,15 @@
 #include "game_sa\Enums\eCarMission.h"
 #include "game_sa\Enums\eClothesModelPart.h"
 #include "game_sa\Enums\eClothesTexturePart.h"
-#include "game_sa\Enums\eCopType.h"
-#include "game_sa\Enums\eCrimeType.h"
 #include "game_sa\Enums\eDecisionMakerEvents.h"
 #include "game_sa\Enums\eEmergencyPedVoices.h"
 #include "game_sa\Enums\eEntityStatus.h"
 #include "game_sa\Enums\eEntityType.h"
 #include "game_sa\Enums\eEventType.h"
-#include "game_sa\Enums\eGangID.h"
 #include "game_sa\Enums\eModelID.h"
-#include "game_sa\Enums\ePadButtons.h"
-#include "game_sa\Enums\ePedAnims.h"
 #include "game_sa\Enums\ePedBones.h"
 #include "game_sa\Enums\ePedModel.h"
-#include "game_sa\Enums\ePedPieceTypes.h"
 #include "game_sa\Enums\ePedState.h"
-#include "game_sa\Enums\ePedType.h"
 #include "game_sa\Enums\eRadioID.h"
 #include "game_sa\Enums\eScriptCommands.h"
 #include "game_sa\Enums\eSoundID.h"
@@ -409,7 +402,6 @@
 #include "game_sa\Enums\eWeaponModel.h"
 #include "game_sa\Enums\eWeaponType.h"
 #include "game_sa\Enums\eWinchType.h"
-#include "game_sa\Enums\ePathType.h"
 #include "game_sa\Fx\CCarFXRenderer.h"
 #include "game_sa\Fx\FxBox_c.h"
 #include "game_sa\Fx\FxEmitterBP_c.h"
