+++ source/game_sa/Entity/Ped/CPed.h
@@ -7,7 +7,6 @@
 #pragma once
 
 #include "PluginBase.h"
-#include "ePedType.h"
 #include "CPhysical.h"
 #include "CAEPedAudioEntity.h"
 #include "CAEPedSpeechAudioEntity.h"
@@ -24,6 +23,52 @@
 #include "CAnimBlendAssociation.h"
 #include "CFire.h"
 #include "CPedGroups.h"
+
+enum ePedType : unsigned int {
+    PED_TYPE_PLAYER1 = 0,
+    PED_TYPE_PLAYER2,
+    PED_TYPE_PLAYER_NETWORK,
+    PED_TYPE_PLAYER_UNUSED,
+    PED_TYPE_CIVMALE,
+    PED_TYPE_CIVFEMALE,
+    PED_TYPE_COP,
+    PED_TYPE_GANG1,
+    PED_TYPE_GANG2,
+    PED_TYPE_GANG3,
+    PED_TYPE_GANG4,
+    PED_TYPE_GANG5,
+    PED_TYPE_GANG6,
+    PED_TYPE_GANG7,
+    PED_TYPE_GANG8,
+    PED_TYPE_GANG9,
+    PED_TYPE_GANG10,
+    PED_TYPE_DEALER,
+    PED_TYPE_MEDIC,
+    PED_TYPE_FIREMAN,
+    PED_TYPE_CRIMINAL,
+    PED_TYPE_BUM,
+    PED_TYPE_PROSTITUTE,
+    PED_TYPE_SPECIAL,
+    PED_TYPE_MISSION1,
+    PED_TYPE_MISSION2,
+    PED_TYPE_MISSION3,
+    PED_TYPE_MISSION4,
+    PED_TYPE_MISSION5,
+    PED_TYPE_MISSION6,
+    PED_TYPE_MISSION7,
+    PED_TYPE_MISSION8
+};
+
+enum ePedPieceTypes {
+    PED_PIECE_UNKNOWN = 0,
+    PED_PIECE_TORSO = 3,
+    PED_PIECE_ASS = 4,
+    PED_PIECE_LEFT_ARM = 5,
+    PED_PIECE_RIGHT_ARM = 6,
+    PED_PIECE_LEFT_LEG = 7,
+    PED_PIECE_RIGHT_LEG = 8,
+    PED_PIECE_HEAD = 9
+};
 
 enum ePedCreatedBy
 {
