--- source/game_sa/Enums/eCopType.h
@@ -1,18 +1,0 @@
-/*
-    Plugin-SDK (Grand Theft Auto San Andreas) header file
-    Authors: GTA Community. See more here
-    https://github.com/DK22Pac/plugin-sdk
-    Do not delete this comment block. Respect others' work!
-*/
-#pragma once
-
-enum eCopType : unsigned int
-{
-	COP_TYPE_CITYCOP,
-	COP_TYPE_LAPDM1,
-	COP_TYPE_SWAT1,
-	COP_TYPE_SWAT2,
-	COP_TYPE_FBI,
-	COP_TYPE_ARMY,
-	COP_TYPE_CSHER = 7
-};
