--- source/game_sa/Enums/eCrimeType.h
@@ -1,26 +1,0 @@
-/*
-    Plugin-SDK (Grand Theft Auto San Andreas) header file
-    Authors: GTA Community. See more here
-    https://github.com/DK22Pac/plugin-sdk
-    Do not delete this comment block. Respect others' work!
-*/
-#pragma once
-
-enum class eCrimeType : std::int32_t  {
-    CRIME_FIRE_WEAPON = 1,
-    CRIME_DAMAGED_PED = 2,
-    CRIME_DAMAGED_COP = 3,
-    CRIME_DAMAGE_CAR = 4,
-    CRIME_DAMAGE_COP_CAR = 5,
-    CRIME_CAR_STEAL = 6,
-    CRIME_VEHICLE_DAMAGE = 8,
-    CRIME_KILL_PED_WITH_CAR = 10,
-    CRIME_KILL_COP_PED_WITH_CAR = 11,
-    CRIME_SET_PED_ON_FIRE = 13,
-    CRIME_SET_COP_PED_ON_FIRE = 14,
-    CRIME_SET_CAR_ON_FIRE = 15,
-    CRIME_EXPLOSION = 17,
-    CRIME_SEALTH_KILL_PED_WITH_KNIFE = 18,
-    CRIME_HIT_CAR = 21,
-    CRIME_AIM_GUN = 22
-};
