--- source/game_sa/Enums/eGangID.h
@@ -1,20 +1,0 @@
-/*
-    Plugin-SDK (Grand Theft Auto San Andreas) header file
-    Authors: GTA Community. See more here
-    https://github.com/DK22Pac/plugin-sdk
-    Do not delete this comment block. Respect others' work!
-*/
-#pragma once
-
-enum  eGangID {
-    GANG_BALLAS = 0,
-    GANG_GROVE = 0x1,
-    GANG_VAGOS = 0x2,
-    GANG_RIFA = 0x3,
-    GANG_DANANGBOYS = 0x4,
-    GANG_MAFIA = 0x5,
-    GANG_TRIAD = 0x6,
-    GANG_AZTECAS = 0x7,
-    GANG_UNUSED1 = 0x8, //!< RUSSIAN_MAFIA
-    GANG_UNUSED2 = 0x9 //!< BIKERS
-};
