--- source/game_sa/Enums/ePathType.h
@@ -1,7 +1,0 @@
-#pragma once
-
-enum ePathType
-{
-    PATH_TYPE_CARS = 0,
-    PATH_TYPE_BOATS 
-};
