--- source/game_sa/Enums/ePedAnims.h
@@ -1,202 +1,0 @@
-/*
-    Plugin-SDK (Grand Theft Auto San Andreas) header file
-    Authors: GTA Community. See more here
-    https://github.com/DK22Pac/plugin-sdk
-    Do not delete this comment block. Respect others' work!
-*/
-#pragma once
-
-enum ePedAnims
-{
-  PED_ANIM_WALK_CIVI = 0x0,
-  PED_ANIM_RUN_CIVI = 0x1,
-  PED_ANIM_SPRINT_PANIC = 0x2,
-  PED_ANIM_IDLE_STANCE = 0x3,
-  PED_ANIM_ROADCROSS = 0x4,
-  PED_ANIM_WALK_START = 0x5,
-  PED_ANIM_RUN_STOP = 0x6,
-  PED_ANIM_RUN_STOPR = 0x7,
-  PED_ANIM_IDLE_HBHB = 0x8,
-  PED_ANIM_IDLE_HBHB_2 = 0x9,
-  PED_ANIM_IDLE_TIRED = 0xA,
-  PED_ANIM_IDLE_ARMED = 0xB,
-  PED_ANIM_IDLE_CHAT = 0xC,
-  PED_ANIM_IDLE_TAXI = 0xD,
-  PED_ANIM_SWIM_TREAD = 0xE,
-  PED_ANIM_KO_SHOT_FRONT = 0xF,
-  PED_ANIM_KO_SHOT_FRONT_2 = 0x10,
-  PED_ANIM_KO_SHOT_FRONT_3 = 0x11,
-  PED_ANIM_KO_SHOT_FRONT_4 = 0x12,
-  PED_ANIM_KO_SHOT_FACE = 0x13,
-  PED_ANIM_KO_SHOT_STOM = 0x14,
-  PED_ANIM_GAS_CWR = 0x15,
-  PED_ANIM_KD_LEFT = 0x16,
-  PED_ANIM_KD_RIGHT = 0x17,
-  PED_ANIM_KO_SKID_FRONT = 0x18,
-  PED_ANIM_KO_SPIN_R = 0x19,
-  PED_ANIM_KO_SKID_BACK = 0x1A,
-  PED_ANIM_KO_SPIN_L = 0x1B,
-  PED_ANIM_SHOT_PARTIAL = 0x1C,
-  PED_ANIM_SHOT_LEFTP = 0x1D,
-  PED_ANIM_SHOT_PARTIAL_B = 0x1E,
-  PED_ANIM_SHOT_RIGHTP = 0x1F,
-  PED_ANIM_HIT_FRONT = 0x20,
-  PED_ANIM_HIT_L = 0x21,
-  PED_ANIM_HIT_BACK = 0x22,
-  PED_ANIM_HIT_R = 0x23,
-  PED_ANIM_FLOOR_HIT = 0x24,
-  PED_ANIM_HIT_WALK = 0x25,
-  PED_ANIM_HIT_WALL = 0x26,
-  PED_ANIM_FLOOR_HIT_F = 0x27,
-  PED_ANIM_HIT_BEHIND = 0x28,
-  PED_ANIM_FIGHTSH_FWD = 0x29,
-  PED_ANIM_FIGHTSH_LEFT = 0x2A,
-  PED_ANIM_FIGHTSH_BWD = 0x2B,
-  PED_ANIM_FIGHTSH_RIGHT = 0x2C,
-  PED_ANIM_FIGHTSHF = 0x2D,
-  PED_ANIM_FIGHTSHB = 0x2E,
-  PED_ANIM_FIGHT2IDLE = 0x2F,
-  PED_ANIM_BOMBER = 0x30,
-  PED_ANIM_GUN_STAND = 0x31,
-  PED_ANIM_GUNMOVE_FWD = 0x32,
-  PED_ANIM_GUNMOVE_L = 0x33,
-  PED_ANIM_GUNMOVE_BWD = 0x34,
-  PED_ANIM_GUNMOVE_R = 0x35,
-  PED_ANIM_GUN_2_IDLE = 0x36,
-  PED_ANIM_WEAPON_CROUCH = 0x37,
-  PED_ANIM_GUNCROUCHFWD = 0x38,
-  PED_ANIM_CROUCH_ROLL_L = 0x39,
-  PED_ANIM_GUNCROUCHBWD = 0x3A,
-  PED_ANIM_CROUCH_ROLL_R = 0x3B,
-  PED_ANIM_CAR_SIT = 0x3C,
-  PED_ANIM_CAR_LSIT = 0x3D,
-  PED_ANIM_CAR_SIT_WEAK = 0x3E,
-  PED_ANIM_CAR_SIT_PRO = 0x3F,
-  PED_ANIM_CAR_SITP = 0x40,
-  PED_ANIM_CAR_SITPLO = 0x41,
-  PED_ANIM_DRIVE_L = 0x42,
-  PED_ANIM_DRIVE_R = 0x43,
-  PED_ANIM_DRIVE_LO_L = 0x44,
-  PED_ANIM_DRIVE_LO_R = 0x45,
-  PED_ANIM_DRIVE_L_WEAK = 0x46,
-  PED_ANIM_DRIVE_R_WEAK = 0x47,
-  PED_ANIM_DRIVE_L_PRO = 0x48,
-  PED_ANIM_DRIVE_R_PRO = 0x49,
-  PED_ANIM_DRIVEBY_L = 0x4A,
-  PED_ANIM_DRIVEBY_R = 0x4B,
-  PED_ANIM_DRIVEBYL_L = 0x4C,
-  PED_ANIM_DRIVEBYL_R = 0x4D,
-  PED_ANIM_CAR_LB = 0x4E,
-  PED_ANIM_CAR_LB_WEAK = 0x4F,
-  PED_ANIM_CAR_LB_PRO = 0x50,
-  PED_ANIM_DRIVE_BOAT = 0x51,
-  PED_ANIM_DRIVE_BOAT_L = 0x52,
-  PED_ANIM_DRIVE_BOAT_R = 0x53,
-  PED_ANIM_DRIVE_BOAT_BACK = 0x54,
-  PED_ANIM_DRIVE_L_SLOW = 0x55,
-  PED_ANIM_DRIVE_R_SLOW = 0x56,
-  PED_ANIM_DRIVE_L_WEAK_SLOW = 0x57,
-  PED_ANIM_DRIVE_R_WEAK_SLOW = 0x58,
-  PED_ANIM_DRIVE_L_PRO_SLOW = 0x59,
-  PED_ANIM_DRIVE_R_PRO_SLOW = 0x5A,
-  PED_ANIM_DRIVE_TRUCK = 0x5B,
-  PED_ANIM_DRIVE_TRUCK_L = 0x5C,
-  PED_ANIM_DRIVE_TRUCK_R = 0x5D,
-  PED_ANIM_DRIVE_TRUCK_BACK = 0x5E,
-  PED_ANIM_KART_DRIVE = 0x5F,
-  PED_ANIM_KART_L = 0x60,
-  PED_ANIM_KART_R = 0x61,
-  PED_ANIM_KART_LB = 0x62,
-  PED_ANIM_BIKE_PICKUPR = 0x63,
-  PED_ANIM_BIKE_PICKUPL = 0x64,
-  PED_ANIM_BIKE_PULLUPR = 0x65,
-  PED_ANIM_BIKE_PULLUPL = 0x66,
-  PED_ANIM_BIKE_ELBOWL = 0x67,
-  PED_ANIM_BIKE_ELBOWR = 0x68,
-  PED_ANIM_BIKE_FALL_OFF = 0x69,
-  PED_ANIM_BIKE_FALLR = 0x6A,
-  PED_ANIM_CAR_HOOKERTALK = 0x6B,
-  PED_ANIM_CAR_CRAWLOUTRHS = 0x6C,
-  PED_ANIM_CAR_CRAWLOUTRHS_2 = 0x6D,
-  PED_ANIM_CAR_ROLLOUT_LHS = 0x6E,
-  PED_ANIM_CAR_ROLLOUT_RHS = 0x6F,
-  PED_ANIM_GETUP = 0x70,
-  PED_ANIM_GETUP_2 = 0x71,
-  PED_ANIM_GETUP_3 = 0x72,
-  PED_ANIM_GETUP_FRONT = 0x73,
-  PED_ANIM_JUMP_LAUNCH = 0x74,
-  PED_ANIM_JUMP_LAUNCH_R = 0x75,
-  PED_ANIM_JUMP_GLIDE = 0x76,
-  PED_ANIM_JUMP_LAND = 0x77,
-  PED_ANIM_FALL_FALL = 0x78,
-  PED_ANIM_FALL_GLIDE = 0x79,
-  PED_ANIM_FALL_LAND = 0x7A,
-  PED_ANIM_FALL_COLLAPSE = 0x7B,
-  PED_ANIM_FALL_BACK = 0x7C,
-  PED_ANIM_FALL_FRONT = 0x7D,
-  PED_ANIM_EV_STEP = 0x7E,
-  PED_ANIM_EV_DIVE = 0x7F,
-  PED_ANIM_CLIMB_JUMP = 0x80,
-  PED_ANIM_CLIMB_IDLE = 0x81,
-  PED_ANIM_CLIMB_PULL = 0x82,
-  PED_ANIM_CLIMB_STAND = 0x83,
-  PED_ANIM_CLIMB_STAND_FINISH = 0x84,
-  PED_ANIM_CLIMB_JUMP_B = 0x85,
-  PED_ANIM_CLIMB_JUMP2FALL = 0x86,
-  PED_ANIM_XPRESSSCRATCH = 0x87,
-  PED_ANIM_TURN_180 = 0x88,
-  PED_ANIM_TURN_L = 0x89,
-  PED_ANIM_TURN_R = 0x8A,
-  PED_ANIM_ARRESTGUN = 0x8B,
-  PED_ANIM_DROWN = 0x8C,
-  PED_ANIM_DUCK_COWER = 0x8D,
-  PED_ANIM_HANDSUP = 0x8E,
-  PED_ANIM_HANDSCOWER = 0x8F,
-  PED_ANIM_FUCKU = 0x90,
-  PED_ANIM_PHONE_IN = 0x91,
-  PED_ANIM_PHONE_OUT = 0x92,
-  PED_ANIM_PHONE_TALK = 0x93,
-  PED_ANIM_SEAT_DOWN = 0x94,
-  PED_ANIM_SEAT_UP = 0x95,
-  PED_ANIM_SEAT_IDLE = 0x96,
-  PED_ANIM_ATM = 0x97,
-  PED_ANIM_ABSEIL = 0x98,
-  PED_ANIM_WALK_DOORPARTIAL = 0x99,
-  PED_ANIM_FACSURP = 0x9A,
-  PED_ANIM_FACSURPM = 0x9B,
-  PED_ANIM_FACURIOS = 0x9C,
-  PED_ANIM_FACANGER = 0x9D,
-  PED_ANIM_FACANGER_2 = 0x9E,
-  PED_ANIM_FACANGER_3 = 0x9F,
-  PED_ANIM_FACTALK = 0xA0,
-  PED_ANIM_FACGUM = 0xA1,
-  PED_ANIM_TAP_HAND = 0xA2,
-  PED_ANIM_TAP_HANDP = 0xA3,
-  PED_ANIM_SHOVE_PARTIAL = 0xA4,
-  PED_ANIM_FLEE_LKAROUND_01 = 0xA5,
-  PED_ANIM_ENDCHAT_01 = 0xA6,
-  PED_ANIM_ENDCHAT_02 = 0xA7,
-  PED_ANIM_ENDCHAT_03 = 0xA8,
-  PED_ANIM_SMOKE_IN_CAR = 0xA9,
-  PED_ANIM_PASS_SMOKE_IN_CAR = 0xAA,
-  PED_ANIM_DAM_ARML_FRMBK = 0xAB,
-  PED_ANIM_DAM_ARML_FRMFT = 0xAC,
-  PED_ANIM_DAM_ARML_FRMLT = 0xAD,
-  PED_ANIM_DAM_ARMR_FRMBK = 0xAE,
-  PED_ANIM_DAM_ARMR_FRMFT = 0xAF,
-  PED_ANIM_DAM_ARMR_FRMRT = 0xB0,
-  PED_ANIM_DAM_LEGL_FRMBK = 0xB1,
-  PED_ANIM_DAM_LEGL_FRMFT = 0xB2,
-  PED_ANIM_DAM_LEGL_FRMLT = 0xB3,
-  PED_ANIM_DAM_LEGR_FRMBK = 0xB4,
-  PED_ANIM_DAM_LEGR_FRMFT = 0xB5,
-  PED_ANIM_DAM_LEGR_FRMRT = 0xB6,
-  PED_ANIM_DAM_STOMACH_FRMBK = 0xB7,
-  PED_ANIM_DAM_STOMACH_FRMFT = 0xB8,
-  PED_ANIM_DAM_STOMACH_FRMLT = 0xB9,
-  PED_ANIM_DAM_STOMACH_FRMRT = 0xBA,
-  PED_ANIM_CAR_DEAD_LHS = 0xBB,
-  PED_ANIM_CAR_DEAD_RHS = 0xBC,
-  PED_ANIM_CAR_TUNE_RADIO = 0xBD,
-  PED_ANIM_GANG_GUNSTAND = 0xBE,
-};
