--- source/game_sa/Enums/ePedPieceTypes.h
@@ -1,14 +1,0 @@
-#pragma once
-
-enum ePedPieceTypes {
-    PED_PIECE_UNKNOWN = 0,
-
-    PED_PIECE_TORSO = 3,
-    PED_PIECE_ASS = 4,
-    PED_PIECE_LEFT_ARM = 5,
-    PED_PIECE_RIGHT_ARM = 6,
-    PED_PIECE_LEFT_LEG = 7,
-    PED_PIECE_RIGHT_LEG = 8,
-    PED_PIECE_HEAD = 9
-};
-
